requirements.txt
Flask==2.0.2
python-dotenv==0.19.2
pydantic==1.10.8
langchain
langchain_pinecone
langchain[docarray]
requests
langchain-core
langchain_openai
openai
langchain_community
langchain_experimental
